src/items/requirements/default.txt
# Do not include azure-functions-worker in this file. The Python Worker is managed by the Azure Functions platform. Manually managing azure-functions-worker may cause unexpected issues.

azure-functions==1.12.0
fastapi==0.89.1
opencensus-ext-logging==0.1.1
opencensus-extension-azure-functions==1.0.0
opencensus==0.11.1
